requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b2
click==8.0.1
Django==3.2.3
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
env==0.1.0
environs==9.3.2
flake8==3.9.2
marshmallow==3.12.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
